moved filteraddbar to a different file and changed contact to be used in both chat and contacts
