+++ WChat/src/app/screens/groups/components/Group/index.js
@@ -1,9 +1,8 @@
-import React, { Component } from 'react';
+import React from 'react';
 import {
   StyleSheet,
   Text,
-  View,
-  Image
+  View
 } from 'react-native';
 import styles from './styles';
