+++ WChat/src/app/screens/groups/index.js
@@ -1,17 +1,17 @@
-import React, { Component } from 'react';
+import React from 'react';
 import groups from '../../../../resources/groups.json';
 import Group from './components/Group';
+import FilterAddBar from '../../components/FilterAddBar';
 import {
   View,
   FlatList,
-  ListItem,
-  Text,
   StyleSheet
 } from 'react-native';
 import styles from './styles'
 
 const GroupList = () => (
   <View style={styles.container}>
+    <FilterAddBar/>
     <FlatList
       data={groups}
       renderItem={({item}) => <Group group={item}/>}
